pyproject.toml
[project]
name = "trisigma" 
version = "1.0.0"
description = "Algo-trading SDK"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE.txt"}
authors = [
  {name = "Arda Gok", email = "ardagkmhs@gmail.com" }
]
maintainers = [
  {name = "Arda Gok", email = "ardagkmhs@gmail.com" }
]

dependencies = [
    "numpy==1.22.3",
    "pandas==1.4.3",
    "requests==2.28.1",
    "cachetools==5.2.0",
    "exchange-calendars==4.2.4",
    "python-dotenv==0.21.0",
    "timeago==1.0.16",
    "toml==0.10.2",
    "pymongo==4.2.0",
    "pymemcache==4.0.0",
    "websocket-client[optional]",
    "python-socks==2.3.0",
    "websocket-client==1.5.1",
    "websockets==10.3",
    "ib-insync==0.9.86",
    "bunny-storm==1.0.28",
    "binance-connector==3.3.0",
    "webull==0.6.1",
    "ibapi==9.81.1.post1"
]
[build-system]
requires = ["setuptools>=67.2.0"]
build-backend = "setuptools.build_meta"
